Worlds of Journalism sample data

A subset of data from the Worlds of Journalism 2012-16 study containing survey data of 1,200 journalists from five European countries.

WoJ

Format

A data frame with 1200 rows and 15 variables:

country

Country of residence

reach

Reach of medium

employment

Current employment situation

temp_contract

Type of contract (if current employment situation is either full-time or part-time

autonomy_selection

Autonomy in news story selection, scale from 1 (no freedom at all) to 5 (complete freedom)

autonomy_emphasis

Autonomy in news story emphasis, scale from 1 (no freedom at all) to 5 (complete freedom)

ethics_1

Agreement with statement "Journalists should always adhere to codes of professional ethics, regardless of situation and context", scale from 1 (strongly disagree) to 5 (strongly agree) (reverse-coded!)

ethics_2

Agreement with statement "What is ethical in journalism depends on the specific situation.", scale from 1 (strongly disagree) to 5 (strongly agree)

ethics_3

Agreement with statement "What is ethical in journalism is a matter of personal judgment.", scale from 1 (strongly disagree) to 5 (strongly agree)

ethics_4

Agreement with statement "It is acceptable to set aside moral standards if extraordinary circumstances require it.", scale from 1 (strongly disagree) to 5 (strongly agree)

work_experience

Work experience as a journalist in years

trust_parliament

Trust placed in parliament, scale from 1 (no trust at all) to 5 (complete trust)

trust_government

Trust placed in government, scale from 1 (no trust at all) to 5 (complete trust)

trust_parties

Trust placed in parties, scale from 1 (no trust at all) to 5 (complete trust)

trust_politicians

Trust placed in politicians in general, scale from 1 (no trust at all) to 5 (complete trust)

Source

https://worldsofjournalism.org/data/data-and-key-tables-2012-2016