Describe numeric variables

Describe numeric variables by several measures of central tendency and variability. If no variables are specified, all numeric (integer or double) variables are described.

describe(data, ..., na.rm = TRUE)

Arguments

data
a tibble

...
Variables to describe (column names). Leave empty to describe all numeric variables in data.

na.rm
a logical value indicating whether NA values should be stripped before the computation proceeds. Defaults to TRUE.

Value

a tibble

See also

Examples

iris %>% describe()

#> # A tibble: 4 x 15
#>   Variable         N Missing     M    SD   Min   Q25   Mdn   Q75   Max Range
#>   <chr>        <int>   <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1 Sepal.Length   150       0  5.84 0.828   4.3   5.1  5.8    6.4   7.9   3.6
#> 2 Sepal.Width    150       0  3.06 0.436   2     2.8  3      3.3   4.4   2.4
#> 3 Petal.Length   150       0  3.76 1.77    1     1.6  4.35   5.1   6.9   5.9
#> 4 Petal.Width    150       0  1.20 0.762   0.1   0.3  1.3    1.8   2.5   2.4
#> # ... with 4 more variables: CI_95_LL <dbl>, CI_95_UL <dbl>, Skewness <dbl>,
#> #   Kurtosis <dbl>

mtcars %>% describe(mpg, am, cyl)

#> # A tibble: 3 x 15
#>   Variable     N Missing      M    SD   Min   Q25   Mdn   Q75   Max Range
#>   <chr>    <int>   <int>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1 mpg         32       0 20.1   6.03   10.4  15.4  19.2  22.8  33.9  23.5
#> 2 am          32       0  0.406 0.499   0     0     0     1     1     1  
#> 3 cyl         32       0  6.19  1.79    4     4     6     8     8     4  
#> # ... with 4 more variables: CI_95_LL <dbl>, CI_95_UL <dbl>, Skewness <dbl>,
#> #   Kurtosis <dbl>